add schema and model for contacts
