+++ models/contact.js
@@ -1,0 +1,52 @@
+const { Schema, model } = require("mongoose");
+const Joi = require("joi");
+
+const { handleMongooseError } = require("../utils");
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+      required: true,
+    },
+    phone: {
+      type: String,
+      required: true,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  { versionKey: false, timestamps: false }
+);
+
+contactSchema.post("save", handleMongooseError);
+
+const addSchema = Joi.object({
+  name: Joi.string().required().messages({
+    "any.required": "missing required name field",
+  }),
+  email: Joi.string().required().messages({
+    "any.required": "missing required email field",
+  }),
+  phone: Joi.string().required().messages({
+    "any.required": "missing required phone field",
+  }),
+  favorite: Joi.boolean(),
+});
+
+const schemas = {
+  addSchema,
+};
+
+const Contact = model("contacts", contactSchema);
+
+module.exports = {
+  Contact,
+  schemas,
+};
