--- schemas/contactsSchema.js
@@ -1,17 +1,0 @@
-const Joi = require("joi");
-
-const addSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "any.required": "missing required name field",
-  }),
-  email: Joi.string().required().messages({
-    "any.required": "missing required email field",
-  }),
-  phone: Joi.string().required().messages({
-    "any.required": "missing required phone field",
-  }),
-});
-
-module.exports = {
-  addSchema,
-};
