+++ utils/handleMongooseError.js
@@ -1,0 +1,6 @@
+const handleMongooseError = (err, data, next) => {
+  err.status = 400;
+  next();
+};
+
+module.exports = handleMongooseError;
