+++ utils/index.js
@@ -1,7 +1,9 @@
-const HttpError = require('./HttpError');
-const ctrlWrapper = require('./ctrlWrapper');
+const HttpError = require("./HttpError");
+const ctrlWrapper = require("./ctrlWrapper");
+const handleMongooseError = require("./handleMongooseError");
 
 module.exports = {
-    HttpError,
-    ctrlWrapper
-}
+  HttpError,
+  ctrlWrapper,
+  handleMongooseError,
+};
